I lied - adding in SY and ETR round
I think the row_number() calculation will display the number of observations in a given year for someone. It should cut by observer, and teacher. EG: the first time in a given year that Laz does an observation for Walters the number is 1. The second time is 2, etc. The following year, the first observation is again 1, then 2, etc.