Update v_surveys.so_assignments_long.sql
Adjusting the assignment logic so that both the manager and the direct report have to be "opted in" to the S&O survey. Before, a manager was assigned S&O surveys when their direct report was giving S&O survey feedback to others. Now a manager is assigned to give a direct report feedback when the manager is opted into the survey AND the direct report is giving others feedback